requirements.txt
python-dotenv
langchain
openai
faiss-cpu
streamlit